modul3assignment7/personal-agent-simple/requirements.txt
langchain==0.3.26
langchain-google-genai==2.1.6
langchain-community
sqlalchemy
python-dotenv
